+++ react/EXPERIMENTAL_useTableBulkActions.ts
@@ -1,3 +1,3 @@
-import useTableBulkActions from './components/EXPERIMENTAL_Table/hooks/useTableBulkActions'
-
-export default useTableBulkActions
+export {
+  default,
+} from './components/EXPERIMENTAL_Table/hooks/useTableBulkActions'